Testing Controller with MatcherFuncs
Step 1: Testing.
I added tests for mapping a Controller with MatcherFuncs.  I'm guessing it needs some work (I am at the end of my work day and don't have time to look in to testing messages, but I'm guessing I need to change some strings in the tests), but it's something.  I'll toy around with it a little later and see what happens.