Updated handlerForOptions to accept arguments of type []MatcherFunc
Because of the extra code it would take (and how ugly that extra code would be) to create a varargs slice of type []interface{} that contained all of the values that need to be passed to Map, MapBefore, and MapAfter from within MapController, I thought it might be a better idea to modify handlerForOptions to accept arguments of type []MatcherFunc as well as MatcherFunc.  This allows me to just generate the matcherFuncs array once in MapController, then pass it on as a normal argument to Map.